pyproject.toml
[tool.poetry]
name = "Basic Face Detection"
version = "0.1.0"
description = "Solve the challange Basic Face Detection from Hackattic"
authors = ["guilherme <guilherme.schroederr@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^1.1.2"
python-dotenv = "^0.15.0"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
